Update add/remove neighbors
