+++ iota-client/core/mod.rs
@@ -1,6 +1,5 @@
 //! Core APIs types and builders
 
-mod add_neighbors;
 mod attach_to_tangle;
 mod broadcast_transactions;
 mod check_consistency;
@@ -9,11 +8,9 @@ mod get_balances;
 mod get_inclusion_states;
 mod get_transactions_to_approve;
 mod get_trytes;
-mod remove_neighbors;
 mod store_transactions;
 mod were_addresses_spent_from;
 
-pub use add_neighbors::AddNeighborsBuilder;
 pub use attach_to_tangle::AttachToTangleBuilder;
 pub use broadcast_transactions::BroadcastTransactionsBuilder;
 pub use check_consistency::CheckConsistencyBuilder;
@@ -22,6 +19,5 @@ pub use get_balances::GetBalancesBuilder;
 pub use get_inclusion_states::GetInclusionStatesBuilder;
 pub use get_transactions_to_approve::GetTransactionsToApproveBuilder;
 pub use get_trytes::GetTrytesBuilder;
-pub use remove_neighbors::RemoveNeighborsBuilder;
 pub use store_transactions::StoreTransactionsBuilder;
 pub use were_addresses_spent_from::WereAddressesSpentFromBuilder;
